Combine database settings and connection strings sections
